refactor(preconsignment): remove pre-consignment related code and clean up schema
