--- backend/internal/preconsignment/router.go
@@ -1,151 +1,0 @@
-package preconsignment
-
-import (
-	"encoding/json"
-	"log/slog"
-	"net/http"
-
-	"github.com/OpenNSW/nsw/internal/auth"
-	"github.com/OpenNSW/nsw/utils"
-)
-
-// Router handles HTTP routing for pre-consignment endpoints.
-type Router struct {
-	pcs *Service
-}
-
-// NewRouter creates a new Router.
-func NewRouter(pcs *Service) *Router {
-	return &Router{
-		pcs: pcs,
-	}
-}
-
-// HandleGetTraderPreConsignments handles GET /api/v1/pre-consignments
-// No query params required for traderId - uses traderId from auth context
-// Pagination query params: offset (optional), limit (optional)
-// Response: TraderPreConsignmentsResponseDTO
-func (r *Router) HandleGetTraderPreConsignments(w http.ResponseWriter, req *http.Request) {
-	ctx := req.Context()
-	authCtx := auth.GetAuthContext(ctx)
-	if authCtx == nil || authCtx.User == nil {
-		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	traderID := authCtx.User.ID
-	offset, limit, err := utils.ParsePaginationParams(req)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	templates, err := r.pcs.GetTraderPreConsignments(req.Context(), traderID, offset, limit)
-	if err != nil {
-		slog.Error("failed to retrieve pre-consignment templates", "error", err)
-		http.Error(w, "failed to retrieve pre-consignment templates: "+err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(templates); err != nil {
-		slog.Error("failed to encode response", "error", err)
-		http.Error(w, "failed to encode response", http.StatusInternalServerError)
-		return
-	}
-}
-
-// HandleCreatePreConsignment handles POST /api/v1/pre-consignments
-func (r *Router) HandleCreatePreConsignment(w http.ResponseWriter, req *http.Request) {
-	ctx := req.Context()
-	authCtx := auth.GetAuthContext(ctx)
-	if authCtx == nil || authCtx.User == nil {
-		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	var createReq CreatePreConsignmentDTO
-	if err := json.NewDecoder(req.Body).Decode(&createReq); err != nil {
-		http.Error(w, "invalid request body: "+err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	traderId := authCtx.User.ID
-	// TODO: Initial trader context is nil; services requiring user metadata should fetch it on-demand
-	// from the user profile service rather than relying on preloaded request context.
-	preConsignment, err := r.pcs.InitializePreConsignment(req.Context(), &createReq, traderId, nil)
-	if err != nil {
-		slog.Error("failed to create pre-consignment", "error", err)
-		http.Error(w, "failed to create pre-consignment: "+err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusCreated)
-	if err := json.NewEncoder(w).Encode(preConsignment); err != nil {
-		slog.Error("failed to encode response", "error", err)
-		http.Error(w, "failed to encode response", http.StatusInternalServerError)
-		return
-	}
-}
-
-// HandleGetPreConsignmentsByTraderID handles GET /api/v1/pre-consignments
-// Returns all pre-consignment instances for authenticated trader
-func (r *Router) HandleGetPreConsignmentsByTraderID(w http.ResponseWriter, req *http.Request) {
-	ctx := req.Context()
-	authCtx := auth.GetAuthContext(ctx)
-	if authCtx == nil || authCtx.User == nil {
-		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	traderID := authCtx.User.ID
-	preConsignments, err := r.pcs.GetPreConsignmentsByTraderID(req.Context(), traderID)
-	if err != nil {
-		slog.Error("failed to retrieve pre-consignments", "error", err)
-		http.Error(w, "failed to retrieve pre-consignments: "+err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(preConsignments); err != nil {
-		slog.Error("failed to encode response", "error", err)
-		http.Error(w, "failed to encode response", http.StatusInternalServerError)
-		return
-	}
-}
-
-// HandleGetPreConsignmentByID handles GET /api/v1/pre-consignments/{preConsignmentId}
-func (r *Router) HandleGetPreConsignmentByID(w http.ResponseWriter, req *http.Request) {
-	ctx := req.Context()
-	authCtx := auth.GetAuthContext(ctx)
-	if authCtx == nil || authCtx.User == nil {
-		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	preConsignmentIDStr := req.PathValue("preConsignmentId")
-	if preConsignmentIDStr == "" {
-		http.Error(w, "pre-consignment ID is required", http.StatusBadRequest)
-		return
-	}
-
-	preConsignmentID := preConsignmentIDStr
-
-	preConsignment, err := r.pcs.GetPreConsignmentByID(req.Context(), preConsignmentID)
-	if err != nil {
-		slog.Error("failed to retrieve pre-consignment", "error", err)
-		http.Error(w, "failed to retrieve pre-consignment: "+err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(preConsignment); err != nil {
-		slog.Error("failed to encode response", "error", err)
-		http.Error(w, "failed to encode response", http.StatusInternalServerError)
-		return
-	}
-}
