--- backend/internal/preconsignment/router_test.go
@@ -1,398 +1,0 @@
-package preconsignment
-
-import (
-	"bytes"
-	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/DATA-DOG/go-sqlmock"
-	"github.com/google/uuid"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-	"gorm.io/driver/postgres"
-	"gorm.io/gorm"
-	"gorm.io/gorm/logger"
-
-	"github.com/OpenNSW/nsw/internal/auth"
-	taskManager "github.com/OpenNSW/nsw/internal/task/manager"
-	workflowManagerV1 "github.com/OpenNSW/nsw/internal/workflow/manager"
-	"github.com/OpenNSW/nsw/internal/workflow/model"
-)
-
-type MockTemplateProvider struct {
-	mock.Mock
-}
-
-func (m *MockTemplateProvider) GetWorkflowTemplateByID(ctx context.Context, id string) (*model.WorkflowTemplate, error) {
-	args := m.Called(ctx, id)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*model.WorkflowTemplate), args.Error(1)
-}
-
-func (m *MockTemplateProvider) GetWorkflowTemplateByIDV2(ctx context.Context, id string) (*model.WorkflowTemplateV2, error) {
-	args := m.Called(ctx, id)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*model.WorkflowTemplateV2), args.Error(1)
-}
-
-func (m *MockTemplateProvider) GetWorkflowNodeTemplatesByIDs(ctx context.Context, ids []string) ([]model.WorkflowNodeTemplate, error) {
-	args := m.Called(ctx, ids)
-	return args.Get(0).([]model.WorkflowNodeTemplate), args.Error(1)
-}
-
-func (m *MockTemplateProvider) GetWorkflowNodeTemplateByID(ctx context.Context, id string) (*model.WorkflowNodeTemplate, error) {
-	args := m.Called(ctx, id)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*model.WorkflowNodeTemplate), args.Error(1)
-}
-
-func (m *MockTemplateProvider) GetEndNodeTemplate(ctx context.Context) (*model.WorkflowNodeTemplate, error) {
-	args := m.Called(ctx)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*model.WorkflowNodeTemplate), args.Error(1)
-}
-
-// MockWorkflowManager implements manager.Manager for testing.
-type MockWorkflowManager struct {
-	mock.Mock
-}
-
-func (m *MockWorkflowManager) StartWorkflowInstance(ctx context.Context, tx *gorm.DB, workflowID string, workflowTemplates []model.WorkflowTemplate, globalContext map[string]any, handler workflowManagerV1.WorkflowEventHandler) error {
-	args := m.Called(ctx, tx, workflowID, workflowTemplates, globalContext, handler)
-	return args.Error(0)
-}
-
-func (m *MockWorkflowManager) GetWorkflowInstance(ctx context.Context, workflowID string) (*model.Workflow, error) {
-	args := m.Called(ctx, workflowID)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*model.Workflow), args.Error(1)
-}
-
-func (m *MockWorkflowManager) RegisterTaskHandler(_ workflowManagerV1.TaskInitHandler) error {
-	return nil
-}
-
-func (m *MockWorkflowManager) HandleTaskUpdate(_ context.Context, _ taskManager.WorkflowManagerNotification) error {
-	return nil
-}
-
-func setupRouterTestDB(t *testing.T) (*gorm.DB, sqlmock.Sqlmock) {
-	mockDB, sqlMock, err := sqlmock.New()
-	if err != nil {
-		t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
-	}
-
-	dialector := postgres.New(postgres.Config{
-		Conn:       mockDB,
-		DriverName: "postgres",
-	})
-
-	db, err := gorm.Open(dialector, &gorm.Config{
-		Logger: logger.Default.LogMode(logger.Silent),
-	})
-	if err != nil {
-		t.Fatalf("an error '%s' was not expected when opening a gorm database connection", err)
-	}
-
-	return db, sqlMock
-}
-
-func withAuthContext(ctx context.Context, userID string) context.Context {
-	authCtx := &auth.AuthContext{
-		User: &auth.UserContext{
-			ID: userID,
-		},
-	}
-	return context.WithValue(ctx, auth.AuthContextKey, authCtx)
-}
-
-func TestPreConsignmentRouter_HandleGetPreConsignmentByID(t *testing.T) {
-	db, sqlMock := setupRouterTestDB(t)
-	mockWM := new(MockWorkflowManager)
-	svc := NewService(db, nil, mockWM)
-	r := NewRouter(svc)
-
-	id := uuid.NewString()
-	templateID := uuid.NewString()
-	nodeTemplateID := uuid.NewString()
-
-	sqlMock.MatchExpectationsInOrder(false)
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignments\"").WillReturnRows(sqlmock.NewRows([]string{"id", "pre_consignment_template_id"}).AddRow(id, templateID))
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignment_templates\"").WillReturnRows(sqlmock.NewRows([]string{"id", "name"}).AddRow(templateID, "Template"))
-
-	mockWM.On("GetWorkflowInstance", mock.Anything, id).Return(&model.Workflow{
-		BaseModel: model.BaseModel{ID: id},
-		Status:    model.WorkflowStatusInProgress,
-		WorkflowNodes: []model.WorkflowNode{
-			{
-				BaseModel:              model.BaseModel{ID: uuid.NewString()},
-				WorkflowNodeTemplateID: nodeTemplateID,
-				State:                  model.WorkflowNodeStateReady,
-				WorkflowNodeTemplate: model.WorkflowNodeTemplate{
-					BaseModel: model.BaseModel{ID: nodeTemplateID},
-					Type:      "TEST",
-				},
-			},
-		},
-	}, nil)
-
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/"+id, nil)
-	req.SetPathValue("preConsignmentId", id)
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-	w := httptest.NewRecorder()
-	r.HandleGetPreConsignmentByID(w, req)
-	assert.Equal(t, http.StatusOK, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetTraderPreConsignments(t *testing.T) {
-	db, sqlMock := setupRouterTestDB(t)
-	svc := NewService(db, nil, nil)
-	r := NewRouter(svc)
-
-	traderID := "trader1"
-	sqlMock.MatchExpectationsInOrder(false)
-	sqlMock.ExpectQuery("(?i)SELECT count").WillReturnRows(sqlmock.NewRows([]string{"count"}).AddRow(1))
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignment_templates\"").WillReturnRows(sqlmock.NewRows([]string{"id", "name"}).AddRow(uuid.NewString(), "Template"))
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM .pre_consignments.").WillReturnRows(sqlmock.NewRows([]string{"id", "trader_id", "pre_consignment_template_id"}).AddRow(uuid.NewString(), traderID, uuid.NewString()))
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM .pre_consignment_templates.").WillReturnRows(sqlmock.NewRows([]string{"id", "name"}))
-
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/templates", nil)
-	req = req.WithContext(withAuthContext(req.Context(), traderID))
-	w := httptest.NewRecorder()
-	r.HandleGetTraderPreConsignments(w, req)
-	assert.Equal(t, http.StatusOK, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleCreatePreConsignment(t *testing.T) {
-	db, sqlMock := setupRouterTestDB(t)
-	tp := new(MockTemplateProvider)
-	mockWM := new(MockWorkflowManager)
-	svc := NewService(db, tp, mockWM)
-	r := NewRouter(svc)
-
-	traderID := "trader1"
-	templateID := uuid.NewString()
-	nodeTemplateID := "00000000-0000-0000-0000-000000000001"
-	preConsignmentID := uuid.NewString()
-
-	payload := CreatePreConsignmentDTO{
-		PreConsignmentTemplateID: templateID,
-	}
-	body, _ := json.Marshal(payload)
-
-	tp.On("GetWorkflowTemplateByID", mock.Anything, mock.Anything).Return(&model.WorkflowTemplate{BaseModel: model.BaseModel{ID: templateID}, NodeTemplates: []string{nodeTemplateID}}, nil)
-
-	sqlMock.MatchExpectationsInOrder(false)
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignment_templates\"").WillReturnRows(sqlmock.NewRows([]string{"id", "workflow_template_id", "depends_on"}).AddRow(templateID, uuid.NewString(), []byte("[]")))
-
-	sqlMock.ExpectBegin()
-	sqlMock.ExpectExec("(?i)INSERT INTO \"pre_consignments\"").WillReturnResult(sqlmock.NewResult(1, 1))
-
-	mockWM.On("StartWorkflowInstance", mock.Anything, mock.Anything, mock.AnythingOfType("string"), mock.Anything, mock.Anything, mock.Anything).Return(nil)
-
-	sqlMock.ExpectCommit()
-
-	// Post-commit reloads
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignments\"").WillReturnRows(sqlmock.NewRows([]string{"id", "pre_consignment_template_id"}).AddRow(preConsignmentID, templateID))
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignment_templates\"").WillReturnRows(sqlmock.NewRows([]string{"id", "name"}).AddRow(templateID, "Template"))
-
-	mockWM.On("GetWorkflowInstance", mock.Anything, mock.AnythingOfType("string")).Return(&model.Workflow{
-		BaseModel: model.BaseModel{ID: preConsignmentID},
-		Status:    model.WorkflowStatusInProgress,
-		WorkflowNodes: []model.WorkflowNode{
-			{
-				BaseModel:              model.BaseModel{ID: uuid.NewString()},
-				WorkflowNodeTemplateID: nodeTemplateID,
-				State:                  model.WorkflowNodeStateReady,
-				WorkflowNodeTemplate: model.WorkflowNodeTemplate{
-					BaseModel: model.BaseModel{ID: nodeTemplateID},
-					Type:      "TEST",
-				},
-			},
-		},
-	}, nil)
-
-	req, _ := http.NewRequest("POST", "/api/v1/pre-consignments", bytes.NewBuffer(body))
-	req = req.WithContext(withAuthContext(req.Context(), traderID))
-	w := httptest.NewRecorder()
-	r.HandleCreatePreConsignment(w, req)
-	assert.Equal(t, http.StatusCreated, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleCreatePreConsignment_InvalidPayload(t *testing.T) {
-	db, _ := setupRouterTestDB(t)
-	svc := NewService(db, nil, nil)
-	r := NewRouter(svc)
-
-	req, _ := http.NewRequest("POST", "/api/v1/pre-consignments", bytes.NewBufferString("invalid json"))
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-	w := httptest.NewRecorder()
-	r.HandleCreatePreConsignment(w, req)
-	assert.Equal(t, http.StatusBadRequest, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetTraderPreConsignments_PaginationError(t *testing.T) {
-	db, _ := setupRouterTestDB(t)
-	svc := NewService(db, nil, nil)
-	r := NewRouter(svc)
-
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/templates?limit=invalid", nil)
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-
-	w := httptest.NewRecorder()
-	r.HandleGetTraderPreConsignments(w, req)
-
-	assert.Equal(t, http.StatusBadRequest, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetTraderPreConsignments_ServiceError(t *testing.T) {
-	db, sqlMock := setupRouterTestDB(t)
-	svc := NewService(db, nil, nil)
-	r := NewRouter(svc)
-
-	sqlMock.ExpectQuery("(?i)SELECT count").WillReturnError(fmt.Errorf("db error"))
-
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/templates", nil)
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-	w := httptest.NewRecorder()
-	r.HandleGetTraderPreConsignments(w, req)
-	assert.Equal(t, http.StatusInternalServerError, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetPreConsignmentByID_InvalidID(t *testing.T) {
-	db, _ := setupRouterTestDB(t)
-	r := NewRouter(NewService(db, nil, nil))
-
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/invalid-uuid", nil)
-	req.SetPathValue("preConsignmentId", "invalid-uuid")
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-	w := httptest.NewRecorder()
-	r.HandleGetPreConsignmentByID(w, req)
-	assert.Equal(t, http.StatusInternalServerError, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetPreConsignmentByID_ServiceError(t *testing.T) {
-	db, sqlMock := setupRouterTestDB(t)
-	r := NewRouter(NewService(db, nil, nil))
-
-	id := uuid.NewString()
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignments\"").WillReturnError(fmt.Errorf("db error"))
-
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/"+id, nil)
-	req.SetPathValue("preConsignmentId", id)
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-	w := httptest.NewRecorder()
-	r.HandleGetPreConsignmentByID(w, req)
-	assert.Equal(t, http.StatusInternalServerError, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleCreatePreConsignment_ServiceError(t *testing.T) {
-	db, sqlMock := setupRouterTestDB(t)
-	tp := new(MockTemplateProvider)
-	r := NewRouter(NewService(db, tp, nil))
-
-	templateID := uuid.NewString()
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignment_templates\"").WillReturnError(fmt.Errorf("db error"))
-
-	payload := CreatePreConsignmentDTO{PreConsignmentTemplateID: templateID}
-	body, _ := json.Marshal(payload)
-	req, _ := http.NewRequest("POST", "/api/v1/pre-consignments", bytes.NewBuffer(body))
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-	w := httptest.NewRecorder()
-	r.HandleCreatePreConsignment(w, req)
-	assert.Equal(t, http.StatusInternalServerError, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetPreConsignmentsByTraderID_Success(t *testing.T) {
-	db, sqlMock := setupRouterTestDB(t)
-	r := NewRouter(NewService(db, nil, nil))
-
-	traderID := "trader1"
-	pcID := uuid.NewString()
-	templateID := uuid.NewString()
-
-	sqlMock.MatchExpectationsInOrder(false)
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignments\"").
-		WillReturnRows(sqlmock.NewRows([]string{"id", "trader_id", "state", "pre_consignment_template_id"}).
-			AddRow(pcID, traderID, "IN_PROGRESS", templateID))
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignment_templates\"").
-		WillReturnRows(sqlmock.NewRows([]string{"id", "name"}).AddRow(templateID, "Template"))
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"workflows\"").
-		WillReturnRows(sqlmock.NewRows([]string{"id", "status"}).AddRow(pcID, "IN_PROGRESS"))
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"workflow_nodes\"").
-		WillReturnRows(sqlmock.NewRows([]string{"id", "workflow_id"}))
-
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments", nil)
-	req = req.WithContext(withAuthContext(req.Context(), traderID))
-	w := httptest.NewRecorder()
-	r.HandleGetPreConsignmentsByTraderID(w, req)
-	assert.Equal(t, http.StatusOK, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetPreConsignmentsByTraderID_Unauthorized(t *testing.T) {
-	r := NewRouter(nil)
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments", nil)
-	w := httptest.NewRecorder()
-	r.HandleGetPreConsignmentsByTraderID(w, req)
-	assert.Equal(t, http.StatusUnauthorized, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetPreConsignmentsByTraderID_ServiceError(t *testing.T) {
-	db, sqlMock := setupRouterTestDB(t)
-	r := NewRouter(NewService(db, nil, nil))
-
-	sqlMock.ExpectQuery("(?i)SELECT .* FROM \"pre_consignments\"").WillReturnError(fmt.Errorf("db error"))
-
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments", nil)
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-	w := httptest.NewRecorder()
-	r.HandleGetPreConsignmentsByTraderID(w, req)
-	assert.Equal(t, http.StatusInternalServerError, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetTraderPreConsignments_Unauthorized(t *testing.T) {
-	r := NewRouter(nil)
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/templates", nil)
-	w := httptest.NewRecorder()
-	r.HandleGetTraderPreConsignments(w, req)
-	assert.Equal(t, http.StatusUnauthorized, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleCreatePreConsignment_Unauthorized(t *testing.T) {
-	r := NewRouter(nil)
-	req, _ := http.NewRequest("POST", "/api/v1/pre-consignments", bytes.NewBufferString("{}"))
-	w := httptest.NewRecorder()
-	r.HandleCreatePreConsignment(w, req)
-	assert.Equal(t, http.StatusUnauthorized, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetPreConsignmentByID_Unauthorized(t *testing.T) {
-	r := NewRouter(nil)
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/id", nil)
-	w := httptest.NewRecorder()
-	r.HandleGetPreConsignmentByID(w, req)
-	assert.Equal(t, http.StatusUnauthorized, w.Code)
-}
-
-func TestPreConsignmentRouter_HandleGetPreConsignmentByID_MissingID(t *testing.T) {
-	r := NewRouter(nil)
-	req, _ := http.NewRequest("GET", "/api/v1/pre-consignments/", nil)
-	req = req.WithContext(withAuthContext(req.Context(), "trader1"))
-	w := httptest.NewRecorder()
-	r.HandleGetPreConsignmentByID(w, req)
-	assert.Equal(t, http.StatusBadRequest, w.Code)
-}
