+++ backend/internal/workflow/model/workflow_node.go
@@ -8,10 +8,6 @@ import (
 
 type WorkflowNodeTemplateType = taskPlugin.Type
 
-const (
-	WorkFlowNodeTypeEndNode WorkflowNodeTemplateType = "END_NODE" // Special type for end nodes that don't correspond to a task plugin
-)
-
 type WorkflowNodeState string
 
 const (
@@ -25,12 +21,10 @@ const (
 // WorkflowNodeTemplate represents a template for a workflow node.
 type WorkflowNodeTemplate struct {
 	BaseModel
-	Name                string                   `gorm:"type:varchar(255);column:name;not null" json:"name"`                                          // Human-readable name of the workflow node template
-	Description         string                   `gorm:"type:text;column:description" json:"description"`                                             // Optional description of the workflow node template
-	Type                WorkflowNodeTemplateType `gorm:"type:varchar(50);column:type;not null" json:"type"`                                           // Type of the workflow node
-	Config              json.RawMessage          `gorm:"type:jsonb;column:config;not null;serializer:json" json:"config"`                             // Configuration specific to the workflow node type
-	DependsOn           StringArray              `gorm:"type:jsonb;column:depends_on;not null;serializer:json" json:"depends_on"`                     // Array of workflow node template IDs this node depends on
-	UnlockConfiguration *UnlockConfig            `gorm:"type:jsonb;column:unlock_configuration;serializer:json" json:"unlockConfiguration,omitempty"` // Optional conditional unlock configuration (supports nested AND/OR boolean expressions). If nil, DependsOn uses AND-all logic.
+	Name        string                   `gorm:"type:varchar(255);column:name;not null" json:"name"`              // Human-readable name of the workflow node template
+	Description string                   `gorm:"type:text;column:description" json:"description"`                 // Optional description of the workflow node template
+	Type        WorkflowNodeTemplateType `gorm:"type:varchar(50);column:type;not null" json:"type"`               // Type of the workflow node
+	Config      json.RawMessage          `gorm:"type:jsonb;column:config;not null;serializer:json" json:"config"` // Configuration specific to the workflow node type
 }
 
 func (wnt *WorkflowNodeTemplate) TableName() string {
@@ -40,13 +34,11 @@ func (wnt *WorkflowNodeTemplate) TableName() string {
 // WorkflowNode represents an instance of a workflow node within a workflow.
 type WorkflowNode struct {
 	BaseModel
-	WorkflowID             string            `gorm:"type:text;column:workflow_id;not null" json:"workflowId"`                                     // Reference to the parent Workflow
-	WorkflowNodeTemplateID string            `gorm:"type:text;column:workflow_node_template_id;not null" json:"workflowNodeTemplateId"`           // Reference to the WorkflowNodeTemplate
-	State                  WorkflowNodeState `gorm:"type:varchar(50);column:state;not null" json:"state"`                                         // State of the workflow node
-	ExtendedState          *string           `gorm:"type:text;column:extended_state" json:"extendedState"`                                        // Optional extended state information (e.g., error details)
-	Outcome                *string           `gorm:"type:varchar(100);column:outcome" json:"outcome,omitempty"`                                   // Outcome sub-state when COMPLETED (e.g., APPROVED, REJECTED)
-	DependsOn              StringArray       `gorm:"type:jsonb;column:depends_on;not null;serializer:json" json:"depends_on"`                     // Array of workflow node IDs this node depends on
-	UnlockConfiguration    *UnlockConfig     `gorm:"type:jsonb;column:unlock_configuration;serializer:json" json:"unlockConfiguration,omitempty"` // Resolved instance-level unlock configuration
+	WorkflowID             string            `gorm:"type:text;column:workflow_id;not null" json:"workflowId"`                           // Reference to the parent Workflow
+	WorkflowNodeTemplateID string            `gorm:"type:text;column:workflow_node_template_id;not null" json:"workflowNodeTemplateId"` // Reference to the WorkflowNodeTemplate
+	State                  WorkflowNodeState `gorm:"type:varchar(50);column:state;not null" json:"state"`                               // State of the workflow node
+	ExtendedState          *string           `gorm:"type:text;column:extended_state" json:"extendedState"`                              // Optional extended state information (e.g., error details)
+	Outcome                *string           `gorm:"type:varchar(100);column:outcome" json:"outcome,omitempty"`                         // Outcome sub-state when COMPLETED (e.g., APPROVED, REJECTED)
 
 	// Relationships
 	Workflow             *Workflow            `gorm:"foreignKey:WorkflowID;references:ID" json:"-"`                                // Associated Workflow
@@ -55,15 +47,6 @@ type WorkflowNode struct {
 
 func (wn *WorkflowNode) TableName() string {
 	return "workflow_nodes"
-}
-
-// UpdateWorkflowNodeDTO is used to update the state of a workflow node.
-type UpdateWorkflowNodeDTO struct {
-	WorkflowNodeID      string            `json:"workflowNodeId" binding:"required"` // Workflow Node ID
-	State               WorkflowNodeState `json:"state"`                             // New state of the workflow node
-	AppendGlobalContext map[string]any    `json:"appendGlobalContext,omitempty"`     // Additional global context to append to the consignment (optional)
-	ExtendedState       *string           `json:"extendedState,omitempty"`           // Optional extended state information (e.g., error details)
-	Outcome             *string           `json:"outcome,omitempty"`                 // Outcome sub-state for COMPLETED transitions (e.g., APPROVED, REJECTED)
 }
 
 // WorkflowNodeResponseDTO represents a workflow node in the response.
@@ -75,7 +58,6 @@ type WorkflowNodeResponseDTO struct {
 	State                WorkflowNodeState               `json:"state"`                   // State of the workflow node
 	ExtendedState        *string                         `json:"extendedState,omitempty"` // Optional extended state information (e.g., error details)
 	Outcome              *string                         `json:"outcome,omitempty"`       // Outcome sub-state when COMPLETED
-	DependsOn            []string                        `json:"depends_on"`              // Array of workflow node IDs this node depends on
 }
 
 type WorkflowEdgeResponseDTO struct {
